fix(consensus): UB in the i: state-leaf key encoding — cross-toolchain state_root fork (CONSENSUS-CRITICAL, flagged for review)
Found by the FIRST run of the new second-platform gate (tools/ci_local.sh
on WSL2 Ubuntu/GCC 13): test-state-proof-composite-key failed on Linux
while green on Windows.

Root cause: build_state_leaves() encoded the i: (applied inbound
receipt) leaf key's src_shard as 8 big-endian bytes by shifting a
ShardId (uint32_t) by 32..56 bits — UNDEFINED BEHAVIOR. MSVC folds the
oversized shift to 0 (the intended zero high bytes); x86-64 GCC masks
the shift count mod 32, duplicating the low bytes into the high
positions. Result: the i: leaf keys — and therefore compute_state_root()
— DIVERGED between toolchains whenever an inbound cross-shard receipt
existed. Two nodes built with different compilers would fork on the
first cross-shard block. Invisible for as long as the project built on
exactly one toolchain.

Fix: widen to uint64_t before shifting (chain.cpp:336) — the encoding
every sibling loop already uses. BYTE-INVARIANT vs the shipped MSVC
behavior (a <2^32 value has zero high bytes either way), verified:
state-root determinism + genesis determinism + composite-key tests all
PASS unchanged on Windows; no pin, genesis hash, or state root moves.
Linux now agrees: ci_local gate 167/167 + all guards green on
Linux x86_64.

Audited the whole bug class: every other BE-shift loop in consensus
serialization (block.cpp amount/fee/nonce u64, MergeEvent u32/u64 pairs,
chain.cpp m:/p: builders, main.cpp param sig_msg) uses width-matched
loops — chain.cpp:336 was the sole instance.

Flagged per the consensus-change discipline: no protocol behavior
changes on the shipped MSVC lineage; this makes other compilers CONFORM
to it. Review before external sync.

+++ src/chain/chain.cpp
@@ -333,7 +333,15 @@ std::vector<crypto::MerkleLeaf> Chain::build_state_leaves() const {
         std::vector<uint8_t> key;
         key.reserve(2 + 8 + 32);
         key.push_back('i'); key.push_back(':');
-        for (int i = 7; i >= 0; --i) key.push_back((src >> (8*i)) & 0xff);
+        // src is ShardId (uint32_t) but the key encodes it as 8 big-endian
+        // bytes; shifting a 32-bit value by 32..56 is UNDEFINED BEHAVIOR and
+        // resolves differently per compiler (MSVC folds to 0; x86 GCC masks
+        // the shift count mod 32) — a cross-toolchain state_root divergence,
+        // i.e. a consensus fork. Widen to uint64_t so the encoding is
+        // well-defined and identical everywhere. Byte-invariant vs the prior
+        // MSVC output (a <2^32 value has zero high bytes either way).
+        for (int i = 7; i >= 0; --i)
+            key.push_back((static_cast<uint64_t>(src) >> (8*i)) & 0xff);
         key.insert(key.end(), tx_hash.begin(), tx_hash.end());
         crypto::SHA256Builder b;
         uint8_t marker = 1; b.append(&marker, 1);  // presence marker
